VACMS-20043 Create shared utilities for Mapbox to isolate Facility Locator applications
