+++ src/applications/facility-locator/actions/mapbox/genBBoxFromAddress.js
@@ -1,11 +1,10 @@
 import mbxGeo from '@mapbox/mapbox-sdk/services/geocoding';
-import mapboxClient from '../../components/MapboxClient';
 import {
-  BOUNDING_RADIUS,
-  EXPANDED_BOUNDING_RADIUS,
   MAPBOX_QUERY_TYPES,
   CountriesList,
-} from '../../constants';
+  mapboxClient,
+} from 'platform/utilities/mapbox';
+import { BOUNDING_RADIUS, EXPANDED_BOUNDING_RADIUS } from '../../constants';
 import {
   GEOCODE_STARTED,
   SEARCH_FAILED,
