--- src/applications/facility-locator/components/MapboxClient.js
@@ -1,5 +1,0 @@
-import MapboxClient from '@mapbox/mapbox-sdk';
-import { mapboxToken } from '../utils/mapboxToken';
-
-const mapboxClient = new MapboxClient({ accessToken: mapboxToken });
-export default mapboxClient;
