+++ src/applications/facility-locator/constants/index.js
@@ -155,24 +155,6 @@ export const MapboxInit = {
 };
 
 /**
- * Mapbox api request countries
- */
-
-export const CountriesList = ['us', 'pr', 'ph', 'gu', 'as', 'mp', 'vi'];
-
-/**
- * Mapbox api request types
- */
-
-export const MAPBOX_QUERY_TYPES = [
-  'place',
-  'region',
-  'postcode',
-  'locality',
-  'country',
-];
-
-/**
  * Max search area in miles
  */
 export const MAX_SEARCH_AREA = 500;
