+++ src/applications/facility-locator/containers/FacilitiesMap.jsx
@@ -8,8 +8,8 @@ import vaDebounce from 'platform/utilities/data/debounce';
 import { isEmpty } from 'lodash';
 import { Tab, TabList, TabPanel, Tabs } from 'react-tabs';
 import recordEvent from 'platform/monitoring/record-event';
+import { mapboxToken } from 'platform/utilities/mapbox';
 import { VaAlert } from '@department-of-veterans-affairs/component-library/dist/react-bindings';
-import { mapboxToken } from '../utils/mapboxToken';
 import {
   clearSearchText,
   clearSearchResults,
