--- src/applications/facility-locator/utils/mapbox.js
@@ -1,24 +1,0 @@
-import mbxGeo from '@mapbox/mapbox-sdk/services/geocoding';
-import mapboxClient from '../components/MapboxClient';
-
-const mbxClient = mbxGeo(mapboxClient);
-import { CountriesList, MAPBOX_QUERY_TYPES } from '../constants';
-
-const getFeaturesFromAddress = query => {
-  return new Promise(resolve => {
-    mbxClient
-      .forwardGeocode({
-        countries: CountriesList,
-        types: MAPBOX_QUERY_TYPES,
-        autocomplete: false,
-        query,
-        proximity: 'ip',
-      })
-      .send()
-      .then(features => {
-        resolve(features);
-      });
-  });
-};
-
-export { getFeaturesFromAddress };
