--- src/applications/facility-locator/utils/mapboxToken.js
@@ -1,3 +1,0 @@
-export const mapboxToken =
-  process.env.MAPBOX_TOKEN ||
-  'pk.eyJ1IjoiYWRob2MiLCJhIjoiY2wyZjNwM3dxMDZ4YjNjbzVwbTZ5aWQ1dyJ9.D8TZ1a4WobqcdYLWntXV_w';
