+++ src/applications/static-pages/facilities/FacilityMapWidgetDynamic.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
+import { mapboxToken } from 'platform/utilities/mapbox';
 import { fetchMultiFacility } from './actions';
-import { mapboxToken } from '../../facility-locator/utils/mapboxToken';
 import { buildAddressArray } from '../../facility-locator/utils/facilityAddress';
 import { staticMapURL } from '../../facility-locator/utils/mapHelpers';
