+++ src/applications/static-pages/facilities/vet-center/NearByVALocations.jsx
@@ -1,13 +1,13 @@
 import React, { useState, useEffect, useCallback } from 'react';
 import PropTypes from 'prop-types';
 import { connect, useDispatch } from 'react-redux';
+import { getFeaturesFromAddress } from 'platform/utilities/mapbox';
 import { multiTypeQuery } from '../actions';
 import {
   calculateBoundingBox,
   convertMetersToMiles,
   distancesToNearbyVetCenters,
 } from '../../../facility-locator/utils/facilityDistance';
-import { getFeaturesFromAddress } from '../../../facility-locator/utils/mapbox';
 import buildFacility from './buildFacility';
 import {
   hasAnyMultiData,
